% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coefs}
\alias{coefs}
\title{coefs pour conditionnement}
\format{A data frame with 9 rows and 4 variables:
\describe{
  \item{recettes}{}
  \item{produit}{}
  \item{coef_viande}{}
  \item{qtite_saucisse}{}
}}
\usage{
coefs
}
\description{
Un dataset contenant les coefs de quantité pour chaque conditionnement quand requis
}
\keyword{datasets}
