% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depenses}
\alias{depenses}
\title{Depense}
\format{A data frame with 19 rows and 3 variables:
\describe{
  \item{achat}{}
  \item{achateur}{}
  \item{prix}{}
}}
\usage{
depenses
}
\description{
Un dataset contenant les dépenses commune
}
\keyword{datasets}
