% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gener}
\alias{gener}
\title{Tableau de généralités}
\format{A data frame with 1 rows and 10 variables:
\describe{
  \item{Nombre de cochons}{}
  \item{Poids total (kg)}{}
  \item{Prix au kilo (€/kg)}{}
  \item{Prix total (€)}{}
  \item{Poids de viandes (kg)}{}
  \item{Poids de chair (kg)}{}
  \item{Poids des foies (kg)}{}
  \item{Poids de gras (kg)}{}
  \item{Poids viandes + chair + foies (kg)}{}
  \item{Prix viandes + chair + foies au kilo (€/kg)}{}
}}
\usage{
gener
}
\description{
Un dataset contenant des généralités
}
\keyword{datasets}
