% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price.R
\name{get_price}
\alias{get_price}
\title{get price by person}
\usage{
get_price(svar, unit)
}
\arguments{
\item{svar}{recette}

\item{unit}{soit un tableau avec le prix unitaire, soit le prix total de la recette}
}
\value{
un tableau avec le prix que doit payer chaque entité
}
\description{
get price by person
}
\examples{
library(dplyr)
get_price("chipolatas", unit = recettes \%>\% filter(recettes == "chipolatas") \%>\% pull(prix_total))
}
