% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unitaire.R
\name{get_unitaire}
\alias{get_unitaire}
\title{Fonction pour calculer le prix unitaire quand des coefs sont impliqués}
\usage{
get_unitaire(svar, prix_sauc = 0)
}
\arguments{
\item{svar}{string. Partie du nom du produit}

\item{prix_sauc}{prix d'une saucisse fraiche}
}
\value{
un tableau avec le prix unitaire en focntion du conditionnement
}
\description{
Fonction pour calculer le prix unitaire quand des coefs sont impliqués
}
\examples{
get_unitaire("pate_foie")
}
