% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recettes}
\alias{recettes}
\title{Tableau des recettes}
\format{A data frame with 25 rows and 6 variables:
\describe{
  \item{recettes}{}
  \item{poids_viandes}{}
  \item{poids_foie}{}
  \item{poids_chair}{}
  \item{poids_total}{}
  \item{prix_total}{}
}}
\usage{
recettes
}
\description{
Un dataset contenant la répartition de la viande dans les recettes
}
\keyword{datasets}
