% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{repartition}
\alias{repartition}
\title{Tableau de répartition des produits par entité}
\format{A data frame with 38 rows and 12 variables:
\describe{
  \item{recettes}{}
  \item{produit}{}
  \item{steph_luc}{}
  \item{estelle_marian}{}
  \item{anaelle_matthieu}{}
  \item{laura_thomas}{}
  \item{marion_judicael}{}
  \item{tiffany}{}
  \item{ben_suk}{}
  \item{commun}{}
  \item{dons}{}
  \item{casse}{}
}}
\usage{
repartition
}
\description{
Un dataset contenant la répartition des produits par entité
}
\keyword{datasets}
