% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gener}
\alias{gener}
\title{Tableau de généralités}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 12 columns.}
\usage{
gener
}
\description{
Un dataset contenant des généralités
}
\keyword{datasets}
